dsdgen_service/requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy
python-dotenv
pytz
email_validator
httpx
lxml
requests
beautifulsoup4
pandas
